DoNothing policy now takes horizon into account
When SimpleEvoAgent reaches the end of its genome, it now does *not*
go forward one tick at a time until the horizon limit, but jumps
straight to the end. Ditto for MCTS in rollout.

This dramatically increases the number of rollouts/evaluations each can
make in a given time budget.