Waiting is now in multiples of 5 ticks
The idea here is that agents can learn to wait for previous actions to finish
before they issue the follow-on orders. (E.g. Move from A to B....wait until this is done... then move from B to C)

Also now using local NTBEA module, so removed the temporary override