EventQueue priority changed
Now all MakeDecision actions are executed first in any tick. They
are responsible for geenrating an action that is put onto the queue
(potentially for the same tick), as well as the next MakeDecision to be taken.

This means that we now cope properly with simultaneous actions (for example
at the start of any game); although the execution of actions in the same tick
will be in random order.