Added limit as attribute to City and Route
This will be used to set a maximum for any single force on the Route
and for the capacity of a City.
If capacity is exceeded in a City, then we set the Offensive maximum to be the limit, so that
the surplus force is not useful. In effect it marks the effective force that can be brought to bear in defence.